pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "malloy"
version = "2022.1007-alpha"
description = "Malloy is an experimental language for describing data relationships and transformations"
readme = "README.md"
# authors = [{name="", email=""}]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
keywords = ["malloy"]
dependencies = [
    "duckdb",
    "google-cloud-bigquery"
]
requires-python = ">3.7"

[project.optional-dependencies]
dev = ["pytest", "pip-tools", "pytest-asyncio", "pandas", "toml", "yapf", "twine", "bumpver"]

[project.urls]
Github = "https://github.com/malloydata/malloy-py"

[project.scripts]
malloy = "malloy.__main__:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.bumpver]
current_version = "2022.1007-alpha"
version_pattern = "YYYY.BUILD[-TAG]"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
    'version = "{version}"'
]
"src/malloy/__init__.py" = ["{version}"]

[tool.yapfignore]
ignore_patterns = [
    "**/*.pb2_grpc.py",
    "**/*.pb2.py",
    "**/*.pb2.pyi",
]

[tool.yapf]
based_on_style = "google"
indent_width = 2

[tool.pytest.ini_options]
# addopts = "--log-cli-level 10" #DEBUG
# addopts = "--log-cli-level 20" #INFO
# addopts = "--log-cli-level 30" #WARNING
# addopts = "--log-cli-level 40" #ERROR
# addopts = "--log-cli-level 50" #CRITICAL